refactor: move AuthContext to models
