+++ rpc/helpers.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from fastapi import HTTPException, Request
 
 from rpc.models import RPCRequest
-from server import AuthContext
+from server.models import AuthContext
 
 if TYPE_CHECKING:
   from server.modules.auth_module import AuthModule
